dom/workers/test/test_chromeWorkerJSM.xul

<?xml version="1.0"?>
<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<window title="DOM Worker Threads Test"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="test();">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>

  <script type="application/javascript">
  <![CDATA[

    function test()
    {
      SimpleTest.waitForExplicitFinish();

      var worker;

      function done()
      {
        worker = null;
        SimpleTest.finish();
      }

      function messageCallback(event) {
        is(event.data, "Done", "Correct message");
        done();
      }

      function errorCallback(event) {
        ok(false, "Worker had an error: " + event.message);
        done();
      }

      Components.utils.import("chrome://mochitests/content/chrome/dom/workers/test/WorkerTest.jsm");

      worker = WorkerTest.go(window.location.href, messageCallback,
                             errorCallback);
    }

  ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display:none;"></div>
    <pre id="test"></pre>
  </body>
  <label id="test-result"/>
</window>
